requirements.txt
asgiref==3.7.2 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==4.3.1 ; python_version >= "3.10" and python_version < "4.0"
django==3.2.25 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "4.0"
drf-spectacular==0.27.1 ; python_version >= "3.10" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.10" and python_version < "4.0"
pillow==10.2.0 ; python_version >= "3.10" and python_version < "4.0"
python-decouple==3.8 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.33.0 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.10.0 ; python_version >= "3.10" and python_version < "3.11"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
whitenoise==6.6.0 ; python_version >= "3.10" and python_version < "4.0"
